fix(streaming): prevent crash when starting with no enabled channels
Adds a validation check in `SCPI_StartStreaming` to ensure at least one
ADC channel is enabled before starting to stream data. If no channels
are enabled, the command returns a "Settings conflict" error.

This prevents a TLB exception that occurs when the streaming process is
initiated without any active channels to provide data.

Fixes #132

+++ firmware/src/services/JSON_Encoder.c
@@ -87,7 +87,7 @@ static size_t generateJsonHeader(char *out, size_t buffSize) {
         (unsigned long long)serialNum,
         tickRate);
 
-    if (written < 0 || written >= (int)buffSize) {
+    if (written < 0 || (size_t)written >= buffSize) {
         return 0;
     }
